ostro os: enable "supported recipes" check by default
As discussed in https://github.com/ostroproject/meta-ostro/pull/186,
the expectation is that the check is needed most of the time and
therefore it gets enabled by default in ostro.conf.

Instructions about it get added to building-images.rst (replacing the
obsolete whitelisting via PNWHITELIST) and local.conf. To ensure that
developers notice that change, the local.conf version gets bumped by
one.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>