ostro os: enable "supported recipes" check in local.conf.sample
Setting this in ostro.conf raised the question how strong the default
needs to be so that it still can be overriden in relevant places, and
what those places could be.

Ultimately, this is a developer's choice and Ostro OS is just
recommending to have the check enabled. Therefore setting the
recommended value in our local.conf.sample seems like the best choice.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>