supported-recipes.bbclass: escape + in unsupported recipe report
The output is supposed to be ready for copying into a supported recipe
file, which means that it must be a valid regular expression. That was
not the case for recipes with a + sign in the name, like gtk+3.

We need to escape that (and only that, other valid characters in
recipe names like - are fine).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>