DOC: clarify .dsk and bmaptool use for EFI platforms only
Issue #113 notes user tried using .dsk format and bmaptool
to bitbake an image and create a bootable USB stick for an
Intel Edison board.

This patch updates documentation to clarify
.dsk format and use of bmaptool is for EFI platforms only
and won't work for Intel Edison (or BeagleBone Black).

Also remove references to .vdi images (replaced by use of .ova)
in the building-images as well as booting-and-installation docs.

Signed-off-by: David Kinder <david.b.kinder@intel.com>