Clean up .vdi image options that are now .ova image options
We're no longer creating .vdi images (for VirtualBox) in favor of
using Open Virtualization Archive (.ova) files,  a pre-packaged appliance file
that can be directly imported to VirtualBox and other virtualization applications.

We updated the .vdi to .ova change in the documentation (fixed in PR#194)
and this PR removes the remaining vestiges in configuration files.

Signed-off-by: David Kinder <david.b.kinder@intel.com>