ostro os: remove questionable zip and iotop
Internal review raised some concerns about the upstream maintenance of
these components. Because there are no hard requirements for them and
there are alternatives for interactive use (tar is better for archival
than zip, and atop can show similar information as iotop), the easiest
solution is to remove the components.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>