image-dsk.bbclass: Add support for OS installer
Since EFI blobs contain predefined PARTUUIDs of rootfs and
are signed we have to generate two different EFI blobs
with different PARTUUIDs. Otherwise the kernel installed
on the device will get confused about which partition to
use as rootfs in case the installation USB stick is still
inserted into the device.

Also export partitions data to the file /boot/emmc-partitions-data
which is consumed by the OS installer in order to format eMMC
accordingly and to set the correct PARTUUID to the device's
rootfs partition.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>