java: Move java behind DISTRO_FEATURES feature
Java is a fairly heavyweight package, and it is likely that at least
some users would like to disable it altogether. Currently both mraa and
upm unconditionally depend on Java, which means it will get included in
the image even if the java-jdk packagegroup is not included.

A new distro feature, java, is introduced to gate both the mraa\upm and
java package group dependencies, so that Java can be easily disabled.

Signed-off-by: Erkka Kääriä <erkka.kaaria@intel.com>