ostro-image.bbclass: install mraa/upm bindings
The Java and Python binding packages of mraa and upm were not
installed before because nothing depended on them. The two runtimes
can't depend on them because it is uncertain whether these optional
features are wanted, and the core devkit cannot depend on them because
it is uncertain whether the runtimes are wanted.

At the image level we can decide that based on the IMAGE_FEATURES.

Fixes: IOTOS-1680

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>