image-dsk.bbclass: add installer param to kernel cmdline
Explicitly add "installer" boot parameter to the kernel command line
used by installation media. This way there's no need to detect if
the media is a USB-stick and as such is supposed to run installer.

The initramfs installed onto the target device's internal
storage shall have the installer disabled always.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>