ostro.conf: compiler workarounds for OpenJDK
Recent OpenJDK changes (meta-java 33ddb28) changed the compile flags handling
so that flags needed when building with GCC 6 get passed to all compilers,
whether they need or support the flag or not. This breaks compilation when the
host compiler is too old to recognize -fno-lifetime-dse.

In ostro.conf we currently select GCC 5, so this GCC 6 workaround is not
needed and can be removed. Doing that in ostro.conf has the advantage that
these gcc compiler settings are kept closer together.

Some better solution must be found for OpenJDK and GCC 6 before
Ostro OS can switch to GCC 6. Upstream was notified.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>