CI: simplify auto.conf in eSDK
The auto.conf file is used by the CI system to make changes specific
only to the CI. Many of them are intentionally not enabled for
normal development by default, and setting OS_VERSION so that
it uses a CI variable breaks swupd-image creation in the eSDK
because the variable isn't set.

So instead of using the auto.conf created by the CI system, we
create one specifically for the eSDK. Currently two changes are
done this way:
- MACHINE
- configuring images as "development images"

That we have to do the later without explicit user consent
is unfortunate; the original goal of the mechanism was that
users must confirm that they want to build images which aren't
suitable for production. With the eSDK that is currently
not possible.

Related-To: IOTOS-1587, IOTOS-1613

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>