ostro-image-swupd.bb: install mraa/upm bindings also in swupd images
The image feature check for mraa and upm language bindings did not
work for swupd images, because the content of the "world-dev" bundle
was based on FEATURE_PACKAGES_ variables and not image features. The
result was that FEATURE_PACKAGES_devkit did not include the Java,
Python and JavaScript language binding packages and they were not
getting installed.

Now the content for "world-dev" gets selected both via packages
(BUNDLE_CONTENTS_WORLD) and features (BUNDLE_FEATURES_WORLD). This
approach is simpler and probably wasn't used from the start because
changing bundle features was only added later.

Fixes: ostro-os/#155

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>