meta-swupd: move swupd client config to distro config, part 1
The updated meta-swupd expects these settings to be for the
image(s). Making them in the swupd client is conceptually wrong
because different images may need different settings and/or developers
may want to change them without recompiling swupd-client.

We can set these values at the distro level instead of per-image
because of the use of SWUPD_IMAGE_PN, the value is valid for different
image recipes.

This is the first half of that change. Removing the
ostroprojectorg.key from swupd-client will follow after updating to
the meta-swupd which uses the distro settings ("meta-swupd: move swupd
client config to distro config").

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>